+++ src/main/java/Car/domain/RacingCar.java
@@ -1,37 +1,62 @@
 package Car.domain;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
-public class RacingCar {
-    private static final int RANDOM_NUMBER_BOUND = 10;
-    private static final int MIN_CAR_COUNT = 0;
+public class Cars {
+    private final int MIN_CAR_COUNT = 0;
 
-    private Cars cars;
+    private List<Car> cars;
 
-    public RacingCar(String[] carNames) {
-        cars = new Cars(carNames);
+    public Cars(String[] carNames) {
+        validateCarsSize(carNames);
+        cars = Arrays.stream(carNames)
+                .map(carName -> new Car(carName))
+                .collect(Collectors.toList());
     }
 
-    public RacingCar(String[] carNames, MovingStrategy movingStrategy) {
-        cars = new Cars(carNames, movingStrategy);
+    public Cars(String[] carNames, MovingStrategy movingStrategy) {
+        validateCarsSize(carNames);
+        cars = Arrays.stream(carNames)
+                .map(carName -> new Car(carName, movingStrategy))
+                .collect(Collectors.toList());
+    }
+
+    private void validateCarsSize(String[] carNames) {
+        if (carNames.length <= MIN_CAR_COUNT) {
+            throw new IllegalArgumentException(MIN_CAR_COUNT + "보다 큰 값을 입력해주세요.");
+        }
+    }
+
+    public int getCarsSize(){
+        return cars.size();
     }
 
     public List<Car> getCars() {
-        return cars.getCars();
+        return cars;
     }
 
-    public void gameStart(Random random, int bound) {
-        List<Integer> inputList = new ArrayList<>();
-        int carsSize = cars.getCarsSize();
-
-        for (int i = 0; i < carsSize; i++) {
-            inputList.add(random.nextInt(bound));
-        }
-
-        cars.move(inputList);
-    }
 
     public List<Car> getWinner() {
-        return cars.getWinner();
+        int maxMoveCount = cars.stream()
+                .mapToInt(Car::getMoveCount)
+                .max()
+                .orElse(0);
+
+        return cars.stream()
+                .filter(car -> car.getMoveCount() == maxMoveCount)
+                .collect(Collectors.toList());
+    }
+
+    public void move(List<Integer> inputList) {
+        int carsSize = cars.size();
+
+        if (inputList.size() != carsSize) {
+            throw new IllegalArgumentException("자동차의 갯수와 입력값이 일치하지 않습니다.");
+        }
+
+        for (int i = 0; i < carsSize; i++) {
+            cars.get(i).move(inputList.get(i));
+        }
     }
 }
