+++ src/test/java/Car/CarTest.java
@@ -1,7 +1,6 @@
 package Car;
 
 import Car.domain.Car;
-import Car.domain.MoveCount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -22,8 +21,8 @@ class CarTest {
 
     @Test
     void 자동차초기상태테스트() {
-        int actual = car.getMoveCount();
-        assertThat(actual).isEqualTo(new MoveCount().getMoveCount());
+        int carMoveCount = car.getMoveCount();
+        assertThat(carMoveCount).isEqualTo(0);
     }
 
     @ParameterizedTest
@@ -36,8 +35,8 @@ class CarTest {
     })
     void 자동차움직임테스트(int inputNumber, int moveAnswer) {
         car.move(inputNumber);
-        int actual = car.getMoveCount();
-        assertThat(actual).isEqualTo(new MoveCount(moveAnswer).getMoveCount());
+        int carMoveCount = car.getMoveCount();
+        assertThat(carMoveCount).isEqualTo(moveAnswer);
     }
 
     @ParameterizedTest
